+++ lib/dates.js
@@ -1,0 +1,5 @@
+import { format } from "date-fns";
+
+export function postDate(dateString) {
+  return format(new Date(dateString), "MMMM dd, yyyy");
+}
