--- news/hello.md
@@ -1,21 +1,0 @@
----
-date: 2020-12-10
-title: Hello
-excerpt: This is an excerpt that will be shown in the index view and in the meta.
----
-
-Hi there!
-
-> How are you?
-
-## Dope
-
-```js
-var me = "great";
-```
-
-I enjoy:
-
-- Going to the beach
-- Going to the beach
-- Going to the pool
