+++ package.json
@@ -16,6 +16,7 @@
     "@docsearch/react": "^1.0.0-alpha.14",
     "@tailwindcss/typography": "^0.3.1",
     "autoprefixer": "^10.1.0",
+    "date-fns": "^2.16.1",
     "flareact": "^1.0.0-canary.8",
     "frontmatter-markdown-loader": "^3.6.2",
     "highlight.js": "^10.3.2",
