+++ pages/docs/[slug].js
@@ -1,8 +1,7 @@
 import Layout from "../../components/Layout";
 import Head from "flareact/head";
-import { useEffect, useRef } from "react";
-import { useRouter } from "flareact/router";
 import { getDocs, getDocsManifest } from "../../lib/docs";
+import useMarkdownClicks from "../../lib/use-markdown-clicks";
 
 export async function getEdgeProps({ params }) {
   const { slug } = params;
@@ -27,28 +26,7 @@ export default function Doc({ content, subheadings, manifest }) {
   const title = `Flareact - ${
     content.match(/<h1.*>(.+)<\/h1>/)?.[1] || "Docs"
   }`;
-  const router = useRouter();
-  const container = useRef();
-
-  // Listen for clicks inside documentation, and use the router context to navigate
-  // rather than full page loads.
-  useEffect(() => {
-    function handleClick(e) {
-      e.preventDefault();
-      const { pathname } = new URL(e.target.href);
-
-      router.push("/docs/[slug]", pathname);
-    }
-
-    if (!container) return;
-    var links = [...container.current.querySelectorAll('a[href^="/"]')];
-
-    links.forEach((link) => link.addEventListener("click", handleClick));
-
-    return () => {
-      links.forEach((link) => link.removeEventListener("click", handleClick));
-    };
-  }, []);
+  const { container } = useMarkdownClicks(content);
 
   return (
     <Layout docs={manifest} subheadings={subheadings}>
